requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
python-dotenv
qrcode
fpdf2
